Remove Modals.jsx and create two separate components named CardModal.jsx and ColumnModal.jsx Co-authored-by: csuzukida <chris.suzukida@gmail.com> Co-authored-by: Igniphis <juneparkoh@gmail.com> Co-authored-by: a-t-klein <alxtkln@gmail.com> Co-authored-by: DavidMNorman <DavidMNorman@gmail.com> Co-authored-by: n8ngo <nathanngo8796@gmail.com>
