+++ client/components/modals/ColumnModal.jsx
@@ -1,5 +1,5 @@
-import React, { Component, useState, useEffect } from 'react';
-import Column from './Column.jsx';
+import React from 'react';
+import CardModal from './CardModal';
 
 // Modal for the columns
 function ColumnModal({
@@ -41,10 +41,10 @@ function ColumnModal({
         <input className="modal-column-input" type="text" required placeholder="column name" />
       </form>
       <div className="modal-button-cont">
-        <button className="modal-text-button" onClick={() => saveData()}>
+        <button className="modal-text-button" type="button" onClick={() => saveData()}>
           SAVE
         </button>
-        <button className="modal-text-button" onClick={() => deleteData()}>
+        <button className="modal-text-button" type="button" onClick={() => deleteData()}>
           DELETE
         </button>
       </div>
@@ -54,33 +54,4 @@ function ColumnModal({
   );
 }
 
-// Modal for the card
-function CardModal({ showCardModal, setShowCardModal }) {
-  const addTask = () => {
-    const newCard = document.querySelector('card-modal-input').value;
-    setShowCardModal(!showCardModal);
-  };
-
-  const deleteTask = () => {
-    setShowCardModal(!showCardModal);
-  };
-
-  return (
-    <div className="modal-home">
-      <form className="modal-form">
-        <h1>ADD CARD</h1>
-        <input className="card-modal-input" type="text" required placeholders="add a task" />
-      </form>
-      <div className="modal-button-cont">
-        <button className="modal-text-button" onClick={() => addTask()}>
-          ADD CARD
-        </button>
-        <button className="modal-text-button" onClick={() => deleteTask()}>
-          CANCEL
-        </button>
-      </div>
-    </div>
-  );
-}
-
-export { ColumnModal, CardModal };
+export default ColumnModal;
